updating iOS project to Xcode 5
* updated the GHUnitIOSTestMain.m in the Tests target based on the
  swizzle approach
  - https://github.com/gabriel/gh-unit/issues/96
* updated Scripts/RunTests.sh
  - now sets `CFFIXED_USER_HOME` based on:
    * http://goo.gl/ifXaqf
    * http://www.egeek.me/2012/10/03/jenkins-ghunit-launch-msg-error/
  - disabled all that funky trapping/launchctl stuff
    * i don't know what it is supposed to do
    * it did not work out of the box
    * it never worked on Jenkins CI
* updated the Test target build settings to use ARC - how this was
  working before is a mystery, and not just to me.
    - https://github.com/Yelp/gh-unit/pull/3
    - thanks @bmelts, @bashery
* fixed Text target framework search path problem - was trying to
  import `iPhoneSimulator5.0.sdk framework`.
* changed project deployment target to 5.1.1 - min to allow 64-bit
  slice
* changed project Makefile to build static lib for all architectures

+++ Classes/GHViewTestCase.h
@@ -86,7 +86,7 @@ reason:@"GHVerifyView can only be called from within a GHViewTestCase class"] ra
  usually corresponds to the latest iOS simulator available).
  */
 @interface GHViewTestCase : GHTestCase {
-  NSInteger imageVerifyCount_;
+  short imageVerifyCount_;
 }
 
 /*!
